app/requirements.txt
pip == 25.1.1
bcrypt == 4.3.0
sqlalchemy == 2.0.17
asyncpg ==0.30.0
psycopg == 3.2.9
psycopg-binary == 3.2.9
alembic == 1.16.4
pydantic == 2.10.6
pydantic-settings == 2.10.1
sqlmodel == 0.0.24
starlette == 0.27.0
fastapi == 0.105.0
uvicorn == 0.25.0
pika == 1.3.2
passlib[bcrypt] == 1.7.4
python-jose[cryptography] == 3.3.0
Jinja2 == 3.1.3
python-multipart == 0.0.9
pytest == 7.4.4
aiohttp
